requirements.txt
# Core dependencies
pymongo>=4.0.0
merit-ai>=0.1.0

# Google Gemini API
google-generativeai>=0.1.0

# Visualization dependencies
matplotlib>=3.5.0
seaborn>=0.11.0

# Web server dependencies
flask>=2.0.0
flask-socketio>=5.0.0

# Utility dependencies
python-dotenv>=0.19.0
argparse>=1.4.0
